requirements.txt
Django==4.1.1
djangorestframework==3.13.1
xrpl-py==1.6.0
psycopg2==2.9.3
redis==4.3.4
aiohttp==3.8.1
django-filter==22.1
celery==5.2.7